+++ libraries/extractor/src/main/java/androidx/media3/extractor/mp3/XingFrame.java
@@ -18,6 +18,7 @@ package androidx.media3.extractor.mp3;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.util.ParsableByteArray;
+import androidx.media3.common.util.Util;
 import androidx.media3.extractor.MpegAudioUtil;
 
 /** Representation of a LAME Xing or Info frame. */
@@ -114,5 +115,21 @@ import androidx.media3.extractor.MpegAudioUtil;
 
     return new XingFrame(
         mpegAudioHeader, frameCount, dataSize, tableOfContents, encoderDelay, encoderPadding);
+  }
+
+  /**
+   * Compute the stream duration, in microseconds, represented by this frame. Returns {@link
+   * C#LENGTH_UNSET} if the frame doesn't contain enough information to compute a duration.
+   */
+  // TODO: b/319235116 - Handle encoder delay and padding when calculating duration.
+  public long computeDurationUs() {
+    if (frameCount == C.LENGTH_UNSET || frameCount == 0) {
+      // If the frame count is missing/invalid, the header can't be used to determine the duration.
+      return C.TIME_UNSET;
+    }
+    // Audio requires both a start and end PCM sample, so subtract one from the sample count before
+    // calculating the duration.
+    return Util.sampleCountToDurationUs(
+        (frameCount * header.samplesPerFrame) - 1, header.sampleRate);
   }
 }